cave-saucisson-app/requirements.txt
Flask==2.0.3
Adafruit-BME280==2.1.0
RPi.GPIO==0.7.0
Flask-Cors==3.0.10
requests==2.26.0